+++ CMakeLists.txt
@@ -6,13 +6,38 @@ set(CMAKE_C_COMPILER gcc)
 # target
 project(c-object LANGUAGES C)
 
+# error code target
+set(ERR_SRC err.c)
+set(ERR_TOOLS error-generator)
+
 # c sources
-aux_source_directory(${CMAKE_CURRENT_SOURCE_DIR} C_SOURCES)
+aux_source_directory(${CMAKE_CURRENT_SOURCE_DIR}/lib/src C_SOURCES)
+list(APPEND C_SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/error/src/err.c main.c)
 
 # c includes
-set(C_INCLUDES ${CMAKE_CURRENT_SOURCE_DIR})
+set(C_INCLUDES ${CMAKE_CURRENT_SOURCE_DIR}/lib/inc
+${CMAKE_CURRENT_SOURCE_DIR}/error/ErrorGenerator/common)
 
-# target
+# compile tools
+add_custom_target(${ERR_TOOLS})
+add_custom_target(${ERR_SRC} ALL DEPENDS ${ERR_TOOLS})
+add_custom_command(TARGET ${ERR_TOOLS}
+    COMMAND make DIR_BUILD:=${CMAKE_CURRENT_BINARY_DIR}/${ERR_TOOLS}
+    WORKING_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/error/ErrorGenerator
+)
+
+# generate code
+add_custom_command(TARGET ${ERR_SRC}
+    COMMAND ${CMAKE_CURRENT_BINARY_DIR}/${ERR_TOOLS}/${ERR_TOOLS}
+    -i="${CMAKE_CURRENT_SOURCE_DIR}/error/ErrorGenerator/common/err.h" 
+    -o="${CMAKE_CURRENT_SOURCE_DIR}/error/src/err.c"
+    -input-begin="//error-code-begin" 
+    -input-end="//error-code-end" 
+    -output-begin="//generate-begin" 
+    -output-end="//generate-end"
+)
+
+# executable
 add_executable(${CMAKE_PROJECT_NAME} ${C_SOURCES})
 
 # compile option
