+++ README.md
@@ -51,4 +51,4 @@ C语言如何实现多态？
 - 当构建对象的时候，根据构造函数的不同将虚函数表指针指向不同的虚函数表。
 - 在调用相应的虚函数的时候，只能通过函数指针表进行调用
 
-![avatar](./memory.png)
+![avatar](./resource/memory.png)
