+++ lib/inc/dual_circle_list.h
