+++ lib/inc/vector.h
