+++ lib/src/dual_circle_list.c
