+++ lib/src/object.c
