+++ lib/src/static_queue.c
