+++ lib/src/vector.c
