+++ resource/memory.drawio
